requirements.txt.txt
ultralytics
streamlit
Pillow
numpy
torch
python==3.12
